renderPoint вынесена в view/point.js
